Started implementing the new HUD and inventory.
A bit more Code restructuring
Fixed the Green Apple/Orange fiasco
Moved the vanilla food system to the new TFC system.

+++ TFC_Shared/src/TFC/Core/TFC_Time.java
@@ -207,7 +207,7 @@ public class TFC_Time
     
     public static float getYearRatio()
     {
-    	return daysInYear / 360;
+    	return (float)daysInYear / 360f;
     }
